finished working EXI function calls
previously Slippi wrote to the hardware registers directly. Unfortunately this wouldn't work with Nintendont because Nintendont writes custom patched code over the EXI functions that make it work. In order to deal with this, I decided to rewrite the code to use the game's EXI functions which will call the patched Nintendont functions