add code to preload stadium transformations
stadium transformation times vary based on how long it takes to load the file. this changes the way transformations are loaded. instead of loading the file when the timer is up, the file is loaded when the timer starts, causing transformations to always start within the same amount of time.