desync: freeze deadupfall physics position
Normally, the fighter's position is centered with the camera by directly
updating the model's position with respect to the camera's view matrix.

However, it also updates the fighter's physics position. This becomes a
problem when screen shake is disabled on one client, as it will cause
the physics position of the fighter being screen ko'd to differ slightly.
This only seems to matter in the case of Ice Climbers, because Nana uses
Popo's physics position to determine her next action.