fix bug causing unnecessarily far rollbacks
This might have been a regression with the doubles code but I'm not sure. In effect what would happen if multiple inputs were received that needed to be checked for rollback was that if any one of them (even the newest) had an incorrect prediction, the logic would roll back all the way to the oldest input no matter what.

This was because the SAVESTATE_FRAME was not being updated until the very end (after all the individual PLAYER_SAVESTATE_FRAMEs had been calculated). This meant that when TRIGGER_ROLBACK was called, it would still use the oldest SAVESTATE_FRAME to decide where to rollback to, instead of the minimum amongst all PLAYER_SAVESTATE_FRAMEs.