move register backup to after free space
this allows for the free space to always start at 0x8 regardless of how many registers are being backed up, should be less error prone.

there is always a chance though that this may cause unforeseen bugs because it's assembly and ASM is brittle. That said, I added additional free space such that every stack frame offset that used to be free space still is.

If a function expected to look at the parent's stack frame, however, by adding the total size of the stack frame (formerly 0xB0), these could have issues. I did this in two places that I could find and fixed both of those.