set init state for stage event codes
prior to this sometimes you could get an event early in the game based on the state of a previous game. this wouldn't have been problematic necessarily but it'd be weird why sometimes you'd get an early event and other times you wouldn't and could make people make bad assumptions when using the events.

I also cleaned up the FoD check for if the platform heights are being initialized to no longer look at the link register.