only free bootloader buffer in alloc process
originally the free was added to fix the GFX artifacts on Nana despawn / food consume but the way it was written caused the game to crash when going directly to change the language to japanese.

I also added a place to free the new scene buffers on scene exit instead of where it was before though I don't know if that's completely necessary.